Update MRC08.md
I fixed the following grammatical errors, typos, and improper use of vocabulary.

- Holders of hte MOR token gain tokens in projects building on Morpheus without having to sell their MOR. 

- The project creates a straight forward path to rewarding Coders and thus attract builers to maintain / expand its Code base.

- Projects get access to Compute (as MOR holders have access to Inference on Morpheus) since the project is holding MOR in the Protocol Owned Liquidity this will poweer their Smart Agent interactions.

- **Daily** - emissions are distributed based on daily capital deposits. The primary benefit of this methodology is that it continually encourages and rewards new depositors. However, there are multiple downsides. First, capital deposits won't be smooth - 10 stETH could be deposited on one day and 100 stETH the next. As such, the Community Capital Front End emissions would be greatly disproportionate to the capital contributed as a full day’s emissions would be provided for 10 stETH one day and 100 stETH the next. Second, this could result in a manipulation of the system. Since it would be limited to a one day time horizon, users could repeatedly deposit to earn the Community Capital Front End emissions, then withdraw funds and deposit a day or two later to earn more again.

- **Cumulative** - emissions are distributed based on the cumulative capital deposits to date. This methodology distributes emissions based on the value of capital provided, regardless if that capital was contributed on a particular day, or months/years in the past, so long as it remains deposited. The downside to this approach is that it could discourage builders to develop new front ends. If a large amount of capital is already allocated, then incremental amounts will be relatively insignificant - meaning builders are less incentivized to continue building for an ever smaller piece of the pie.

- **1 Year Duration** - capital deposits through a front-end are relevant for one year so long as they remain deposited. Emissions are distributed based on the pro rata share of the trailing one year deposits. The length of the timeline could be set to any number of days/years. The goal would be to find the appropriate balance which encourages new front end development to solicit capital while removing unnecessary transactions and potential exploitation.

- **Relative Movements across Assets** - Currencies will move relative to each other after being deposited. For example, if token A is deposited when its price is $100 and Token B is deposited when its price is $200, then at the time of deposit, it is a 1:2 ratio. A user who deposited 1 Token A will be credited with half the value of a user who deposited 1 Token B. However, if a week later, Token A’s price has stayed at $100 while Token B’s price has increased to $300, the ratio is now at 1:3 for those same deposits. Prices will need to be updated on a regular basis to ensure that the user’s deposited value represents actual prices. 

The below example shows, using random assets and numbers for presentational purposes only, the blended pro-rata approach where a 50% weight is applied to the TVL and a 50% weight is applied to the daily yield.

## Why is this important? 
Without a protocol-level reward, a community member would be required to form traditional entities and other means of coordinating efforts. This would cause friction within the construct of Morpheus’ use of the Techno Capital Machine; it's critical that the rewards be able to cover the full stack of developers involved.

The question becomes defining the metrics. It’s worth the extra work to define these rewards. Consider Lido the creator of stETH, which has a “reward sharing program” that has driven much of its success in onboarding large amounts of ETH onto its protocol.  

The Web3 community has long suffered from a lack of front-end development. This is mostly due to protocols only offering a reward for proof of work (Compute) or proof of Stake (Capital), thus there are normally no resources directed toward those building at the application layer. The Morpheus network is unique in that it rewards developers of the Protocol, Network, Smart Contracts, and also Smart Agents, Websites & Dapps the Community uses. In this way Morpheus is a truly full-stack project, extending its rewards beyond the back-end functions, right up to front ends that serve the end users.